+++ pytests/operators/test_join.py
@@ -2,7 +2,7 @@ from typing import List, Optional, Tuple
 
 import bytewax.operators as op
 from bytewax.dataflow import Dataflow
-from bytewax.operators import JoinMode, _JoinState
+from bytewax.operators import JoinEmitMode, JoinInsertMode, _JoinState
 from bytewax.testing import TestingSink, TestingSource, run_main
 
 
@@ -25,28 +25,39 @@ def _build_join_dataflow(
     inp_l: List[int],
     inp_r: List[int],
     out: List[Tuple[Optional[int], Optional[int]]],
-    mode: Optional[JoinMode] = None,
+    insert_mode: Optional[JoinInsertMode] = None,
+    emit_mode: Optional[JoinEmitMode] = None,
 ) -> Dataflow:
     flow = Dataflow("test_df")
     lefts = op.input("inp_l", flow, TestingSource(inp_l))
     keyed_lefts = op.key_on("key_l", lefts, lambda _: "ALL")
     rights = op.input("inp_r", flow, TestingSource(inp_r))
     keyed_rights = op.key_on("key_r", rights, lambda _: "ALL")
-    if mode is not None:
-        joined = op.join("join", keyed_lefts, keyed_rights, mode=mode)
+    if insert_mode is not None and emit_mode is not None:
+        joined = op.join(
+            "join",
+            keyed_lefts,
+            keyed_rights,
+            insert_mode=insert_mode,
+            emit_mode=emit_mode,
+        )
     else:
-        joined = op.join("join", keyed_lefts, keyed_rights)
+        joined = op.join(
+            "join",
+            keyed_lefts,
+            keyed_rights,
+        )
     unkeyed = op.key_rm("unkey", joined)
     op.output("out", unkeyed, TestingSink(out))
     return flow
 
 
-def test_join_complete() -> None:
+def test_join_last_complete() -> None:
     inp_l = [1]
     inp_r = [2]
     out: List[Tuple[Optional[int], Optional[int]]] = []
 
-    flow = _build_join_dataflow(inp_l, inp_r, out, "complete")
+    flow = _build_join_dataflow(inp_l, inp_r, out, "last", "complete")
 
     run_main(flow)
     assert out == [
@@ -54,7 +65,7 @@ def test_join_complete() -> None:
     ]
 
 
-def test_join_default_is_complete() -> None:
+def test_join_default_is_last_complete() -> None:
     inp_l = [1]
     inp_r = [2]
     out: List[Tuple[Optional[int], Optional[int]]] = []
@@ -67,12 +78,12 @@ def test_join_default_is_complete() -> None:
     ]
 
 
-def test_join_final() -> None:
+def test_join_last_final() -> None:
     inp_l = [1]
     inp_r = [2, 3]
     out: List[Tuple[Optional[int], Optional[int]]] = []
 
-    flow = _build_join_dataflow(inp_l, inp_r, out, "final")
+    flow = _build_join_dataflow(inp_l, inp_r, out, "last", "final")
 
     run_main(flow)
     assert out == [
@@ -80,12 +91,12 @@ def test_join_final() -> None:
     ]
 
 
-def test_join_running() -> None:
+def test_join_last_running() -> None:
     inp_l = [1]
     inp_r = [2, 3]
     out: List[Tuple[Optional[int], Optional[int]]] = []
 
-    flow = _build_join_dataflow(inp_l, inp_r, out, "running")
+    flow = _build_join_dataflow(inp_l, inp_r, out, "last", "running")
 
     run_main(flow)
     assert out == [
@@ -95,12 +106,12 @@ def test_join_running() -> None:
     ]
 
 
-def test_join_product() -> None:
+def test_join_product_final() -> None:
     inp_l = [1, 2]
     inp_r = [3, 4]
     out: List[Tuple[Optional[int], Optional[int]]] = []
 
-    flow = _build_join_dataflow(inp_l, inp_r, out, "product")
+    flow = _build_join_dataflow(inp_l, inp_r, out, "product", "final")
 
     run_main(flow)
     assert out == [